Revert "Add triangle_multi_complete mode"
